+++ doc/changelog.md
@@ -1,8 +1,10 @@
 ## Changelog
 
 ### 1.4.6
-- #86 android fallback button is now hidden if AllowAlternativeAuthentication is false
+- #51 allows setting the `DefaultColor` to change the color of the fingerprint icon
+- #86 android fallback button is now hidden if AllowAlternativeAuthentication is false (thnaks @fedemkr)
 - #91 new API `GetAuthenticationTypeAsync` to retrieve the biometric auth type
+- CrossFingerprint is now setable for mocking during unit tests (thanks @ArtjomP)
 
 #### 1.4.6-beta4
 - #84 missing resource ids should be available again